YT-26145: Revert visible changes for config serialization
<Message NOT for release notes>

* Changelog entry
  Type: fix
  Component: cpp-sdk

Revert breaking public changes in header files
commit_hash:e25a66243042b5961984505e6ecd67cb9f9f2635
