+++ yt/cpp/mapreduce/interface/config.h
@@ -8,7 +8,6 @@
 #include <library/cpp/yt/yson_string/public.h>
 
 #include <library/cpp/yson/node/node.h>
-#include <library/cpp/yson/node/serialize.h>
 
 #include <util/generic/maybe.h>
 #include <util/generic/string.h>
@@ -18,6 +17,20 @@
 #include <util/datetime/base.h>
 
 namespace NYT {
+
+////////////////////////////////////////////////////////////////////////////////
+
+namespace NYson {
+
+////////////////////////////////////////////////////////////////////////////////
+
+struct IYsonConsumer;
+
+enum class EYsonFormat : int;
+
+////////////////////////////////////////////////////////////////////////////////
+
+} // namespace NYson
 
 ////////////////////////////////////////////////////////////////////////////////
 
@@ -61,19 +74,18 @@ enum class ETraceHttpRequestsMode
     Always /* "always" */,
 };
 
-enum class EUploadDeduplicationMode : int
-{
+DEFINE_ENUM(EUploadDeduplicationMode,
     // For each file only one process' thread from all possible hosts can upload it to the file cache at the same time.
     // The others will wait for the uploading to finish and use already cached file.
-    Global,
+    ((Global)   (0))
 
     // For each file and each particular host only one process' thread can upload it to the file cache at the same time.
     // The others will wait for the uploading to finish and use already cached file.
-    Host,
+    ((Host)     (1))
 
     // All processes' threads will upload a file to the cache concurrently.
-    Disabled,
-};
+    ((Disabled) (2))
+);
 
 ////////////////////////////////////////////////////////////////////////////////
